ch569w-evt : maintain a minimum main.c with LED flashing, all others moved to test.c
